// src/enums/actions.ts
export enum EventsActionTypeEnum {
  GET_ALL_EVENTS = 'GET_ALL_EVENTS',
  UPDATE_EVENT_VIEW = 'UPDATE_EVENT_VIEW',
  UPDATE_EVENT_SCORE = 'UPDATE_EVENT_SCORE',
  UPDATE_EVENT_BY_INDEX = 'UPDATE_EVENT_BY_INDEX',
  UPDATE_EVENT_STREAM_URL = 'UPDATE_EVENT_STREAM_URL',
  GET_DETAIL_EVENT = 'GET_DETAIL_EVENT',
  GET_OTHER_EVENTS = 'GET_OTHER_EVENTS',
  PUBLISH_NEW_EVENT = 'PUBLISH_NEW_EVENT',
  UPDATE_EVENT_PROOF = 'UPDATE_EVENT_PROOF',
  GET_ALL_REPORTS = 'GET_ALL_REPORTS',
  UPDATE_EVENT = 'UPDATE_EVENT',
  UPDATE_EXPAND = 'UPDATE_EXPAND',
  RESET = 'RESET',
}